+++ lib/hooks.js
@@ -28,6 +28,7 @@ const {
   FST_ERR_HOOK_TIMEOUT,
   FST_ERR_HOOK_NOT_SUPPORTED,
   AVVIO_ERRORS_MAP,
+  FST_ERR_HOOK_INVALID_ASYNC_HANDLER,
   appendStackTrace
 } = require('./errors')
 
@@ -408,6 +409,52 @@ function hookIterator (fn, request, reply, next) {
   return fn(request, reply, next)
 }
 
+// wrapper that we expose to the user for hooks handling
+function addHook (name, fn) {
+  if (fn == null) {
+    throw new FST_ERR_HOOK_INVALID_HANDLER(name, fn)
+  }
+
+  if (name === 'onSend' || name === 'preSerialization' || name === 'onError' || name === 'preParsing') {
+    if (fn.constructor.name === 'AsyncFunction' && fn.length === 4) {
+      throw new FST_ERR_HOOK_INVALID_ASYNC_HANDLER()
+    }
+  } else if (name === 'onReady' || name === 'onListen') {
+    if (fn.constructor.name === 'AsyncFunction' && fn.length !== 0) {
+      throw new FST_ERR_HOOK_INVALID_ASYNC_HANDLER()
+    }
+  } else if (name === 'onRequestAbort') {
+    if (fn.constructor.name === 'AsyncFunction' && fn.length !== 1) {
+      throw new FST_ERR_HOOK_INVALID_ASYNC_HANDLER()
+    }
+  } else {
+    if (fn.constructor.name === 'AsyncFunction' && fn.length === 3) {
+      throw new FST_ERR_HOOK_INVALID_ASYNC_HANDLER()
+    }
+  }
+
+  if (name === 'onClose') {
+    this.onClose(fn.bind(this))
+  } else if (name === 'onReady' || name === 'onListen' || name === 'onRoute') {
+    this[kHooks].add(name, fn)
+  } else {
+    this.after((err, done) => {
+      try {
+        _addHook.call(this, name, fn)
+        done(err)
+      } catch (err) {
+        done(err)
+      }
+    })
+  }
+  return this
+
+  function _addHook (name, fn) {
+    this[kHooks].add(name, fn)
+    this[kChildren].forEach(child => _addHook.call(child, name, fn))
+  }
+}
+
 module.exports = {
   Hooks,
   buildHooks,
@@ -425,5 +472,6 @@ module.exports = {
   onRequestHookRunner,
   onTimeoutHookRunner,
   lifecycleHooks,
-  supportedHooks
+  supportedHooks,
+  addHook
 }
